+++ Modules/Intent.Modules.EntityFrameworkCore/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.0.35
+
+- Improvement: Improved installation process when multiple persistence technologies are present.
+
 ### Version 5.0.34
 
 - Improvement: Updated NuGet package versions.